Poltergeist/requirements.txt
Flask==3.0.2
nltk==3.8.1
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==21.2.0
scikit-learn==1.5.2
Werkzeug==3.0.1 